Initial Port of Android -> Typescript
