+++ .gitignore
@@ -1,1 +1,3 @@
 .idea
+node_modules/**
+bluetooth.android.test.*
