+++ _references.d.ts
@@ -1,0 +1,89 @@
+/// <reference path="node_modules/tns-core-modules/ui/builder/builder.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/repeater/repeater.d.ts" />
+/// <reference path="node_modules/tns-core-modules/es-collections.d.ts" />
+/// <reference path="node_modules/tns-core-modules/es6-promise.d.ts" />
+/// <reference path="node_modules/tns-core-modules/declarations.d.ts" />
+/// <reference path="node_modules/tns-core-modules/application/application.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/frame/frame.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/page/page.d.ts" />
+/// <reference path="android17.d.ts" />
+/// <reference path="node_modules/tns-core-modules/console/console.d.ts" />
+/// <reference path="node_modules/tns-core-modules/declarations.d.ts" />
+/// <reference path="node_modules/tns-platform-declarations/org.nativescript.widgets.d.ts" />
+/// <reference path="node_modules/tns-core-modules/color/color.d.ts" />
+/// <reference path="node_modules/tns-core-modules/camera/camera.d.ts" />
+/// <reference path="node_modules/tns-platform-declarations/ios.d.ts" />
+/// <reference path="node_modules/tns-core-modules/data/virtual-array/virtual-array.d.ts" />
+/// <reference path="node_modules/tns-core-modules/data/observable-array/observable-array.d.ts" />
+/// <reference path="node_modules/tns-core-modules/file-system/file-name-resolver.d.ts" />
+/// <reference path="node_modules/tns-core-modules/data/observable/observable.d.ts" />
+/// <reference path="node_modules/tns-core-modules/application-settings/application-settings.d.ts" />
+/// <reference path="node_modules/tns-core-modules/file-system/file-system.d.ts" />
+/// <reference path="node_modules/tns-core-modules/file-system/file-system-access.d.ts" />
+/// <reference path="node_modules/tns-core-modules/fps-meter/fps-meter.d.ts" />
+/// <reference path="node_modules/tns-core-modules/fps-meter/fps-native.d.ts" />
+/// <reference path="node_modules/tns-core-modules/http/http.d.ts" />
+/// <reference path="node_modules/tns-core-modules/image-source/image-source.d.ts" />
+/// <reference path="node_modules/tns-core-modules/js-libs/esprima/esprima.d.ts" />
+/// <reference path="node_modules/tns-core-modules/location/location.d.ts" />
+/// <reference path="node_modules/tns-core-modules/platform/platform.d.ts" />
+/// <reference path="node_modules/tns-core-modules/text/text.d.ts" />
+/// <reference path="node_modules/tns-core-modules/text/span.d.ts" />
+/// <reference path="node_modules/tns-core-modules/text/formatted-string.d.ts" />
+/// <reference path="node_modules/tns-core-modules/timer/timer.d.ts" />
+/// <reference path="node_modules/tns-core-modules/trace/trace.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/activity-indicator/activity-indicator.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/border/border.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/button/button.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/content-view/content-view.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/core/bindable.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/core/control-state-change.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/core/dependency-observable.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/core/proxy.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/core/view.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/core/weak-event-listener.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/date-picker/date-picker.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/dialogs/dialogs.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/editable-text-base/editable-text-base.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/enums/enums.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/gestures/gestures.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/image/image.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/image-cache/image-cache.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/label/label.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/layouts/absolute-layout/absolute-layout.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/layouts/dock-layout/dock-layout.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/layouts/grid-layout/grid-layout.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/layouts/stack-layout/stack-layout.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/layouts/wrap-layout/wrap-layout.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/layouts/layout.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/list-picker/list-picker.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/list-view/list-view.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/placeholder/placeholder.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/progress/progress.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/scroll-view/scroll-view.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/search-bar/search-bar.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/segmented-bar/segmented-bar.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/slider/slider.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/styling/css-selector.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/styling/style-property.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/switch/switch.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/styling/styling.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/styling/visual-state-constants.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/tab-view/tab-view.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/text-base/text-base.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/text-field/text-field.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/text-view/text-view.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/time-picker/time-picker.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/web-view/web-view.d.ts" />
+/// <reference path="node_modules/tns-core-modules/utils/types.d.ts" />
+/// <reference path="node_modules/tns-core-modules/utils/utils.d.ts" />
+/// <reference path="node_modules/tns-core-modules/xml/xml.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/action-bar/action-bar.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/html-view/html-view.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/styling/background.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/styling/font.d.ts" />
+/// <reference path="node_modules/tns-core-modules/css/reworkcss.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/layouts/layout-base.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/animation/animation.d.ts" />
+/// <reference path="node_modules/tns-core-modules/module.d.ts" />
+/// <reference path="node_modules/tns-core-modules/ui/transition/transition.d.ts" />
