+++ tsconfig.json
@@ -1,0 +1,18 @@
+{
+    "compilerOptions": {
+        "noEmitOnError": true,
+        "noEmitHelpers": true,
+        "noLib": false,
+        "sourceMap": false,
+        "target": "es5",
+        "module": "commonjs",
+        "outDir": "bin/dist",
+        "rootDir": "./", 
+        "removeComments": false
+    }, 
+    "exclude": [
+        "node_modules", 
+        "sample", 
+        "bin"
+    ]
+}
